test(within): share proptest strategies, drop duplicate/misplaced tests
