+++ crates/within/src/operator/tests.rs
@@ -1,7 +1,3 @@
-// ===========================================================================
-// design tests
-// ===========================================================================
-
 mod design_tests {
     use crate::domain::Design;
     use crate::observation::FactorMajorStore;
@@ -329,105 +325,6 @@ mod weighted_adjoint_proptests {
         }
     }
 }
-
-// ===========================================================================
-// csr_block tests
-// ===========================================================================
-
-mod csr_block_tests {
-    use crate::csr_block::CsrBlock;
-
-    fn sample_block() -> CsrBlock {
-        // 3x4 matrix:
-        //  [1 0 2 0]
-        //  [0 3 0 4]
-        //  [5 0 0 6]
-        CsrBlock {
-            indptr: vec![0, 2, 4, 6],
-            indices: vec![0, 2, 1, 3, 0, 3],
-            data: vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0],
-            nrows: 3,
-            ncols: 4,
-        }
-    }
-
-    #[test]
-    fn test_nnz() {
-        assert_eq!(sample_block().nnz(), 6);
-    }
-
-    #[test]
-    fn test_transpose_dimensions() {
-        let a = sample_block();
-        let at = a.transpose();
-        assert_eq!(at.nrows, 4);
-        assert_eq!(at.ncols, 3);
-        assert_eq!(at.nnz(), a.nnz());
-    }
-
-    #[test]
-    fn test_transpose_roundtrip() {
-        let a = sample_block();
-        let att = a.transpose().transpose();
-        assert_eq!(att.nrows, a.nrows);
-        assert_eq!(att.ncols, a.ncols);
-        assert_eq!(att.indptr, a.indptr);
-        assert_eq!(att.indices, a.indices);
-        assert_eq!(att.data, a.data);
-    }
-
-    #[test]
-    fn test_transpose_values() {
-        let a = sample_block();
-        let at = a.transpose();
-        // A^T should be 4x3:
-        //  [1 0 5]
-        //  [0 3 0]
-        //  [2 0 0]
-        //  [0 4 6]
-        assert_eq!(at.indptr, vec![0, 2, 3, 4, 6]);
-        assert_eq!(at.indices, vec![0, 2, 1, 0, 1, 2]);
-        assert_eq!(at.data, vec![1.0, 5.0, 3.0, 2.0, 4.0, 6.0]);
-    }
-
-    #[test]
-    fn test_from_dense_table() {
-        let table = vec![1.0, 0.0, 2.0, 0.0, 0.0, 3.0, 0.0, 4.0, 5.0, 0.0, 0.0, 6.0];
-        let a = CsrBlock::from_dense_table(&table, 3, 4);
-        let expected = sample_block();
-        assert_eq!(a.indptr, expected.indptr);
-        assert_eq!(a.indices, expected.indices);
-        assert_eq!(a.data, expected.data);
-        assert_eq!(a.nrows, expected.nrows);
-        assert_eq!(a.ncols, expected.ncols);
-    }
-
-    #[test]
-    fn test_empty_block() {
-        let a = CsrBlock {
-            indptr: vec![0, 0, 0],
-            indices: vec![],
-            data: vec![],
-            nrows: 2,
-            ncols: 3,
-        };
-        assert_eq!(a.nnz(), 0);
-        let at = a.transpose();
-        assert_eq!(at.nrows, 3);
-        assert_eq!(at.ncols, 2);
-        assert_eq!(at.nnz(), 0);
-    }
-
-    #[test]
-    fn test_from_dense_table_all_zeros() {
-        let table = vec![0.0; 6];
-        let a = CsrBlock::from_dense_table(&table, 2, 3);
-        assert_eq!(a.nnz(), 0);
-        assert_eq!(a.indptr, vec![0, 0, 0]);
-    }
-}
-
-// ===========================================================================
 
 // ===========================================================================
 // schwarz tests
