+++ crates/within/tests/common/property_strategies.rs
@@ -1,0 +1,41 @@
+//! Shared proptest strategies for the property-test binaries.
+#![allow(dead_code)]
+
+use ndarray::Array2;
+use proptest::prelude::*;
+use within::PreconditionerConfig;
+
+/// Generate a random FE problem: (categories Array2<u32>, y Vec<f64>).
+/// 2–3 factors, 2–30 levels each, 50–500 observations.
+pub fn random_fe_problem_strategy() -> impl Strategy<Value = (Array2<u32>, Vec<f64>)> {
+    (2..=3u32).prop_flat_map(|n_factors| {
+        let levels = proptest::collection::vec(2..=30u32, n_factors as usize);
+        levels.prop_flat_map(move |n_levels| {
+            let n_obs_range = 50..=500usize;
+            n_obs_range.prop_flat_map(move |n_obs| {
+                let n_levels_clone = n_levels.clone();
+                let cat_cols: Vec<_> = n_levels_clone
+                    .iter()
+                    .map(|&nl| proptest::collection::vec(0..nl, n_obs))
+                    .collect();
+                let y_vec = proptest::collection::vec(-10.0f64..10.0, n_obs);
+                (cat_cols, y_vec).prop_map(move |(cols, y)| {
+                    let n_f = cols.len();
+                    let n = cols[0].len();
+                    let mut cats = Array2::<u32>::zeros((n, n_f));
+                    for (f, col) in cols.iter().enumerate() {
+                        for (i, &val) in col.iter().enumerate() {
+                            cats[[i, f]] = val;
+                        }
+                    }
+                    (cats, y)
+                })
+            })
+        })
+    })
+}
+
+/// Default additive Schwarz preconditioner config for property tests.
+pub fn additive_precond() -> PreconditionerConfig {
+    PreconditionerConfig::default()
+}
