+++ crates/within/tests/orchestrate_solve.rs
@@ -72,22 +72,3 @@ fn test_lsmr_least_squares_weighted_preconditioned() {
     common::assert_converged_with_small_residual(&result, 1e-6);
     common::assert_solution_finite(&result);
 }
-
-#[test]
-fn test_lsmr_weighted() {
-    let design =
-        common::make_design(vec![vec![0, 1, 0, 1, 2], vec![0, 0, 1, 1, 0]]).expect("valid design");
-    let weights = vec![1.0, 2.0, 1.5, 0.5, 3.0];
-    let y = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-
-    let params = LsmrOptions {
-        tol: 1e-8,
-        maxiter: 1000,
-        ..Default::default()
-    };
-    let precond = PreconditionerConfig::default();
-    let solver = Solver::new(design, Some(weights), Some(&precond)).expect("build solver");
-    let result = solver.solve(&y, &params).expect("solve");
-    common::assert_converged_with_small_residual(&result, 1e-6);
-    common::assert_solution_finite(&result);
-}
